Move sql related code to sqlNotebook folder
