+++ src/sql/workbench/services/notebook/sqlNotebook/sqlSessionManager.ts
